+++ generate/adapter_v2_multi_modal.py
@@ -12,7 +12,7 @@ import torch
 from PIL import Image
 
 from finetune.adapter_v2_multi_modal import encode_image
-from lit_llama.adapter_v2_multi_modal import LLaMA
+from lit_llama.adapter_v2_multi_modal import LLaMA, generate
 from lit_llama.tokenizer import Tokenizer
 from lit_llama.utils import lazy_load
 from scripts.prepare_alpaca import generate_prompt
@@ -35,55 +35,6 @@ tokenizer = Tokenizer(tokenizer_path)
 
 clip_model, clip_transform = clip.load("ViT-L/14")
 clip_model = fabric.to_device(clip_model)
-
-
-@torch.no_grad()
-def generate(
-    model: torch.nn.Module,
-    idx: torch.Tensor,
-    max_new_tokens: int,
-    max_seq_length: int,
-    temperature: float = 1.0,
-    top_k = None,
-    eos_id = None,
-    **kwargs,
-) -> torch.Tensor:
-
-    # create an empty tensor of the expected final shape and fill in the current tokens
-    T = idx.size(0)
-    T_new = T + max_new_tokens
-    empty = torch.empty(T_new, dtype=idx.dtype, device=idx.device)
-    empty[:T] = idx
-    idx = empty
-
-    # generate max_new_tokens tokens
-    for t in range(T, T_new):
-        # ignore the not-filled-yet tokens
-        idx_cond = idx[:t]
-        # if the sequence context is growing too long we must crop it at max_seq_length
-        idx_cond = idx_cond if t <= max_seq_length else idx_cond[-max_seq_length:]
-
-        # forward
-        logits = model(idx_cond.view(1, -1), **kwargs)
-        logits = logits[0, -1] / temperature
-
-        # optionally crop the logits to only the top k options
-        if top_k is not None:
-            v, _ = torch.topk(logits, min(top_k, logits.size(-1)))
-            logits[logits < v[[-1]]] = -float("Inf")
-
-        probs = torch.nn.functional.softmax(logits, dim=-1)
-        idx_next = torch.multinomial(probs, num_samples=1)
-
-        # concatenate the new generation
-        # https://github.com/pytorch/pytorch/issues/101936
-        idx[t] = idx_next.item() if idx.device.type == "mps" else idx_next
-
-        # if <eos> token is triggered, return the output (stop generation)
-        if idx_next == eos_id:
-            return idx[:t + 1]  # include the EOS token
-
-    return idx
 
 
 def predict(image, prompt, temperature):
